+++ examples/process_cb.rs
@@ -15,7 +15,8 @@ fn read_frames() -> impl Iterator<Item = Vec<u8>> {
 }
 
 fn process_frames(frames: &[Vec<u8>]) {
-    let mut parser = term_wm_vt100::Parser::new_with_callbacks(24, 80, 0, Callbacks);
+    let mut parser =
+        term_wm_vt100::Parser::new_with_callbacks(24, 80, 0, Callbacks);
     for frame in frames {
         parser.process(frame);
     }